update user and note entities to use i64 for IDs

+++ src/app/handlers/activity_pub/create.rs
@@ -11,7 +11,7 @@ use crate::{
 };
 
 pub async fn get(
-    Path((_, id)): Path<(String, i32)>,
+    Path((_, id)): Path<(String, i64)>,
     State(state): State<AppState>,
 ) -> Result<Response, StatusCode> {
     let note_repository: &dyn NoteRepository = &state.storage;
