+++ src/domain/repositories/user.rs
@@ -5,10 +5,10 @@ use crate::domain::entities::user;
 
 #[async_trait]
 pub trait UserRepository: Send + Sync {
-    async fn find_by_id(&self, id: i32) -> Result<Option<user::Model>, DbErr>;
+    async fn find_by_id(&self, id: i64) -> Result<Option<user::Model>, DbErr>;
     async fn find_by_username(&self, username: &str) -> Result<Option<user::Model>, DbErr>;
     async fn create(&self, username: &str, display_name: &str) -> Result<user::Model, DbErr>;
     async fn update(&self, user: user::ActiveModel) -> Result<user::Model, DbErr>;
-    async fn delete(&self, id: i32) -> Result<(), DbErr>;
+    async fn delete(&self, id: i64) -> Result<(), DbErr>;
     async fn list(&self, limit: u64, offset: u64) -> Result<Vec<user::Model>, DbErr>;
 }
